refactor: folderization components
